Second commit for README.md (lab1)
Molli and Matt both made changes together. We added contribution guidelines, bugs and todo list, inspiration, and license information to readme. The contribution guidelines, the bugs and todo list information is currently vague since the project has not been started yet and the read me is a work in progress. The project logo and screenshots are still to come.